1.20.6 Update
- update dependencies
- moved screen to dedicated class
- fixed the enchantment data fetching
- include license in built JAR
