+++ build.gradle
@@ -1,13 +1,13 @@
 plugins {
     id 'org.ajoberstar.grgit' version '5.0.0'
-    id 'fabric-loom' version '1.4-SNAPSHOT'
+    id 'fabric-loom' version '1.6-SNAPSHOT'
 }
 
-sourceCompatibility = targetCompatibility = JavaVersion.VERSION_17
-
-archivesBaseName = project.archives_base_name
-version = "${project.mod_version}+mc${project.minecraft_version}-${getVersionMetadata()}"
-group = project.maven_group
+base {
+    archivesBaseName = project.archives_base_name
+    version = "${project.mod_version}+mc${project.minecraft_version}-${getVersionMetadata()}"
+    group = project.maven_group
+}
 
 repositories {
     maven {
@@ -34,17 +34,34 @@ dependencies {
     modImplementation "meteordevelopment:meteor-client:${project.meteor_version}"
 }
 
-processResources {
-    inputs.properties("version": project.version, "minecraft_version": project.minecraft_version)
+tasks {
+    processResources {
+        def propertyMap = [
+            "version"   : project.version,
+            "mc_version": project.minecraft_version,
+        ]
 
-    filesMatching("fabric.mod.json") {
-        expand("version": project.version, "minecraft_version": project.minecraft_version)
+        filesMatching("fabric.mod.json") {
+            expand(propertyMap)
+        }
+    }
+
+    jar {
+        from("LICENSE") {
+            rename { "${it}_${project.base.archivesBaseName}" }
+        }
+    }
+
+    java {
+        sourceCompatibility = JavaVersion.VERSION_21
+        targetCompatibility = JavaVersion.VERSION_21
+    }
+
+    tasks.withType(JavaCompile).configureEach {
+        it.options.encoding("UTF-8")
     }
 }
 
-tasks.withType(JavaCompile).configureEach {
-    it.options.encoding("UTF-8")
-}
 
 configurations.modImplementation {
     exclude group: 'club.minnced' // fuck you
