+++ gradle.properties
@@ -1,13 +1,13 @@
 org.gradle.jvmargs=-Xmx4G
 
 # Fabric (https://fabricmc.net/versions.html)
-minecraft_version=1.20.4
-yarn_mappings=1.20.4+build.2
-loader_version=0.15.1
+minecraft_version=1.20.6
+yarn_mappings=1.20.6+build.1
+loader_version=0.15.10
 
 # Mod Properties
 mod_version=1.4.4
 maven_group=maxsuperman.addons
 archives_base_name=villager-roller
 
-meteor_version=0.5.6-SNAPSHOT
+meteor_version=0.5.7-SNAPSHOT
